% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_thistemplate}
\alias{use_thistemplate}
\title{Use thistemplate theme for pkgdown}
\usage{
use_thistemplate(github_url = NULL, author = NULL)
}
\arguments{
\item{github_url}{Optional. The GitHub URL of your package. If provided, will be added to the navbar.}

\item{author}{Optional. The author name. If not provided, will be extracted from DESCRIPTION.}
}
\description{
This function helps you set up the thistemplate theme for your package's pkgdown site.
It will:
\enumerate{
\item Create or update _pkgdown.yml with the necessary template configuration
\item Add thistemplate to the Config/Needs/website field in DESCRIPTION
}
}
