% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mxtemplate.R
\name{use_mxtemplate}
\alias{use_mxtemplate}
\title{Use mxtemplate theme for pkgdown}
\usage{
use_mxtemplate(github_url = NULL)
}
\arguments{
\item{github_url}{Optional. The GitHub URL of your package. If provided, will be added to the navbar.}
}
\description{
This function helps you set up the mxtemplate theme for your package's pkgdown site.
It will:
\enumerate{
\item Create or update _pkgdown.yml with the necessary template configuration
\item Add mxtemplate to the Config/Needs/website field in DESCRIPTION
}
}
